adding the user's crud

+++ app/controllers/user_controller.py
@@ -1,31 +1,36 @@
 from sqlalchemy.orm import Session
-from sqlalchemy.exc import NoResultFound
-from fastapi import Depends
-
+from fastapi import HTTPException, status
 
 from ..models.user_model import User
 from ..schemas.user_schema import UserCreate
 
 
 def get_user(user_id: int, db: Session):
-    try:
-        return db.query(User).filter(User.id == user_id).first()
-    except NoResultFound:
-        return None
+    user = db.query(User).filter(User.id == user_id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
+    return user
 
+def get_users(db: Session):
+    users = db.query(User).all()
+    return users
 
 def get_user_by_email(email: str, db: Session):
-    try:
-        return db.query(User).filter(User.email == email).first()
-    except NoResultFound:
-        return None
+    user = db.query(User).filter(User.email == email).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
+    return user
 
 
 def create_user(user: UserCreate, db: Session):
     db_user = User(
         email=user.email,
         hashed_password=user.hashed_password,
-        permissions=user.permissions
+        permissions=user.permissions,
     )
     db.add(db_user)
     db.commit()
@@ -33,8 +38,26 @@ def create_user(user: UserCreate, db: Session):
     return db_user
 
 
+def update_user(user_id: int, user: UserCreate, db: Session):
+    db_user = db.query(User).filter(User.id == user_id).first()
+    if not db_user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
+    db_user.email = user.email
+    db_user.hashed_password = user.hashed_password
+    db_user.permissions = user.permissions
+    db.commit()
+    db.refresh(db_user)
+    return db_user
+
+
 def delete_user(user_id: int, db: Session):
     db_user = db.query(User).filter(User.id == user_id).first()
+    if not db_user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
+        )
     db.delete(db_user)
     db.commit()
     return db_user
